requirements.txt
pillow
numpy
keras
argparse
GDAL
glob2
tensorflow
openpyxl